third_party/WebKit/LayoutTests/http/tests/devtools/startup/sources/debugger/script-formatter-breakpoints-1-expected.txt

Tests the script formatting is working fine with breakpoints.


Running: testSetup

Running: testBreakpointsInOriginalAndFormattedSource
Script execution paused.
Breakpoint sidebar pane while paused in raw
script-formatter-breakpoints-1.html:10    var f=0;
Script execution resumed.
Script execution paused.
Breakpoint sidebar pane while paused in pretty printed
script-formatter-breakpoints-1.html:formatted:12                var f = 0;
Breakpoint sidebar pane while paused in raw
script-formatter-breakpoints-1.html:10    var f=0;
Script execution resumed.
